cleanup from security review of array and string code
I found a few issues with overflow, and an issue where a TypedArray
helper function was not protecting against the buffer being detached. I
also did some cleanup with inconsistent uses of size_t, charcount_t and
the like.

+++ lib/Runtime/ByteCode/AuxArray.h
@@ -17,7 +17,7 @@ namespace Js
         }
 
         void SetCount(uint count) { this->count = count; }
-        uint32 GetDataSize() const { return sizeof(AuxArray) + sizeof(T) * count; }
+        size_t GetDataSize() const { return sizeof(AuxArray) + sizeof(T) * count; }
     };
     typedef AuxArray<Var> VarArray;
